renaming counter to app
